Initial boilerplate for rustc_wrapper
