+++ cargo-auditable/src/rustc_wrapper.rs
@@ -1,0 +1,3 @@
+pub fn main() {
+    todo!()
+}
